+++ src/app/component-library/ui-elements/ui-elements.module.ts
@@ -2,6 +2,7 @@ import {CommonModule} from '@angular/common';
 import {NgModule} from '@angular/core';
 import {MatIconModule} from '@angular/material/icon';
 import {MatProgressSpinnerModule} from '@angular/material/progress-spinner';
+import {MatTooltipModule} from '@angular/material/tooltip';
 import {AgoComponent} from './ago/ago.component';
 import {AgoPipe} from './ago/ago.pipe';
 import {DateComponent} from './date/date.component';
@@ -18,6 +19,7 @@ import {LoadingStateComponent} from './loading-state/loading-state.component';
     // Angular Material
     MatIconModule,
     MatProgressSpinnerModule,
+    MatTooltipModule,
   ],
   declarations: [
     AgoComponent,