+++ src/app/core/services/api/edc-api.service.ts
@@ -1,19 +1,23 @@
 import {Inject, Injectable} from '@angular/core';
 import {Observable, from} from 'rxjs';
+import {map} from 'rxjs/operators';
 import {
   AssetPage,
   ConnectorLimits,
+  ContractAgreementCard,
   ContractAgreementPage,
   ContractDefinitionPage,
   ContractDefinitionRequest,
   ContractNegotiationRequest,
   DashboardPage,
   EdcClient,
+  GetContractAgreementPageRequest,
   IdResponseDto,
   InitiateCustomTransferRequest,
   InitiateTransferRequest,
   PolicyDefinitionCreateRequest,
   PolicyDefinitionPage,
+  TerminateContractAgreementRequest,
   TransferHistoryPage,
   UiAsset,
   UiAssetCreateRequest,
@@ -23,6 +27,7 @@ import {
   buildEdcClient,
 } from '@sovity.de/edc-client';
 import {APP_CONFIG, AppConfig} from '../../config/app-config';
+import {throwIfNull} from '../../utils/rxjs-utils';
 import {EDC_FAKE_BACKEND} from './fake-backend/edc-fake-backend';
 
 @Injectable({providedIn: 'root'})
@@ -127,6 +132,16 @@ export class EdcApiService {
     );
   }
 
+  terminateContractAgreement(
+    terminateContractAgreementRequest: TerminateContractAgreementRequest,
+  ): Observable<IdResponseDto> {
+    return from(
+      this.edcClient.uiApi.terminateContractAgreement(
+        terminateContractAgreementRequest,
+      ),
+    );
+  }
+
   getContractNegotiation(
     contractNegotiationId: string,
   ): Observable<UiContractNegotiation> {
@@ -135,8 +150,31 @@ export class EdcApiService {
     );
   }
 
-  getContractAgreementPage(): Observable<ContractAgreementPage> {
-    return from(this.edcClient.uiApi.getContractAgreementPage());
+  getContractAgreementPage(
+    getContractAgreementPageRequest: GetContractAgreementPageRequest,
+  ): Observable<ContractAgreementPage> {
+    return from(
+      this.edcClient.uiApi.getContractAgreementPage(
+        getContractAgreementPageRequest,
+      ),
+    );
+  }
+
+  getContractAgreementById(
+    contractAgreementId: string,
+  ): Observable<ContractAgreementCard> {
+    return this.getContractAgreementPage({
+      contractAgreementPageQuery: {terminationStatus: undefined},
+    }).pipe(
+      map((page) =>
+        page.contractAgreements.find(
+          (it) => it.contractAgreementId === contractAgreementId,
+        ),
+      ),
+      throwIfNull(
+        `Contract Agreement with ID ${contractAgreementId} not found`,
+      ),
+    );
   }
 
   initiateTransfer(