+++ src/app/routes/connector-ui/contract-agreement-page/contract-agreement-cards/contract-agreement-card-mapped.service.ts
@@ -19,6 +19,7 @@ export class ContractAgreementCardMappedService {
     contractAgreement: ContractAgreementCard,
   ): ContractAgreementCardMapped {
     const asset = this.assetBuilder.buildAsset(contractAgreement.asset);
+    const isTerminated = contractAgreement.terminationStatus === 'TERMINATED';
 
     return {
       ...contractAgreement,
@@ -27,9 +28,11 @@ export class ContractAgreementCardMappedService {
         (it) => it.state.simplifiedState === 'RUNNING',
       ),
       isConsumingLimitsEnforced: false,
-      statusText: '',
+      isTerminated: isTerminated,
+      showStatus: isTerminated,
+      statusText: isTerminated ? 'Terminated' : '',
       statusTooltipText: '',
-      canTransfer: true,
+      canTransfer: !isTerminated,
       searchTargets: [
         contractAgreement.contractAgreementId,
         contractAgreement.counterPartyId,
@@ -50,16 +53,28 @@ export class ContractAgreementCardMappedService {
     maxConsumingContracts: number,
     agreements: ContractAgreementCardMapped[],
   ): ContractAgreementCardMapped[] {
-    return agreements.map((it, index) => ({
-      ...it,
-      isConsumingLimitsEnforced: true,
-      statusText: index < maxConsumingContracts ? 'Active' : 'Inactive',
-      statusTooltipText: this.getConsumingContractsInfoText(
-        index,
-        maxConsumingContracts,
-      ),
-      canTransfer: index < maxConsumingContracts,
-    }));
+    let activeContractCounter = 0;
+
+    return agreements.map((it) => {
+      if (it.isTerminated) {
+        return it;
+      }
+
+      const modifiedAgreement = {
+        ...it,
+        isConsumingLimitsEnforced: true,
+        showStatus: true,
+        statusText:
+          activeContractCounter < maxConsumingContracts ? 'Active' : 'Inactive',
+        statusTooltipText: this.getConsumingContractsInfoText(
+          activeContractCounter,
+          maxConsumingContracts,
+        ),
+        canTransfer: activeContractCounter < maxConsumingContracts,
+      };
+      activeContractCounter++;
+      return modifiedAgreement;
+    });
   }
 
   private getConsumingContractsInfoText(